% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.helpers.R
\name{bam_summary_cigar}
\alias{bam_summary_cigar}
\title{Summarize CIGAR string by addition}
\usage{
bam_summary_cigar(cigar, which, FUNC = sum)
}
\arguments{
\item{cigar}{CIGAR strings as character vector}

\item{which}{CIGAR operators to extract}

\item{FUNC}{Summarize function. Accept numeric vector and output length = 1.}
}
\value{
Numeric vector of sums of specified CIGAR operators.
}
\description{
Summarize CIGAR string by addition
}
\examples{
cigar.test <- c(
  "148H47M1D113M1D34M1D39M460H",
  "50M"
)
bam_summary_cigar(
  # All Ops that consume reference
  # Therefore sum = reference lengths
  cigar.test, which = c("M", "D", "N", "=", "X")
)
}
