% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.tests.betaReg.R
\name{diff_breg}
\alias{diff_breg}
\title{Differential analysis with beta regression}
\usage{
diff_breg(se, design, target, n.cores = "auto")
}
\arguments{
\item{se}{SummarizedExperiment containing data to test.}

\item{design}{Design formula.}

\item{target}{Which target to perform test.}

\item{n.cores}{Multicore by \code{parallel}. Either "auto", integer, or NA.
NA means no parallel processing.}
}
\value{
data.frame of test results with FDR.
}
\description{
Differential analysis with beta regression
}
\examples{
# TODO
}
