% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.tests.betaReg.R
\name{diff_breg_single}
\alias{diff_breg_single}
\title{Differential analysis with beta regression}
\usage{
diff_breg_single(data, design, target, check = TRUE)
}
\arguments{
\item{data}{Data to use.}

\item{design}{Design formula of the problem.}

\item{target}{Which target to test.}

\item{check}{Whether to perform sanity checks.}
}
\value{
data.frame of test results.
}
\description{
This takes data.frame input.
}
\examples{
# TODO.
}
