% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsAnalysis-package.R
\docType{package}
\name{gsAnalysis-package}
\alias{gsAnalysis}
\alias{gsAnalysis-package}
\title{gsAnalysis: Miscellaneous tools for genomic sequence analysis}
\description{
A miscellaneous toolbox for various genomic sequence analysis tasks. Refer to package vignettes for different topics covered in this package. Part of the y3628 analysis suite.
}
\author{
\strong{Maintainer}: Ye Yuan \email{yeyu@umich.edu} (\href{https://orcid.org/0000-0001-9641-9102}{ORCID})

}
\keyword{internal}
