% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{IRFinderS_read}
\alias{IRFinderS_read}
\title{Read IRFinder-S output of a single sample}
\usage{
IRFinderS_read(result.dir, type.samples = "validated")
}
\arguments{
\item{result.dir}{IRFinder-S output directory}

\item{type.samples}{Which summary type to read in, see details}
}
\value{
tibble of IRFinder-S results
}
\description{
Read IRFinder-S output of a single sample
}
\details{
Two sample types are output by IRFinder-S: \code{validated} and \code{full}.
\itemize{
\item \code{validated}: read the "IRFinder-IR-nondir-val.txt"
\item \code{full}: read the "IRFinder-IR-nondir.txt"
}

For column names please refer to the function body
}
\examples{
#TODO
}
