% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{IRFinderS_readSamples}
\alias{IRFinderS_readSamples}
\title{Read IRFinder-S output of multiple samples}
\usage{
IRFinderS_readSamples(
  named.result.dirs,
  score.column = "IR.ratio",
  join.column = c("chr", "start", "end", "strand", "symbol"),
  type.samples = "validated",
  min.samples = 3,
  wl = 0
)
}
\arguments{
\item{named.result.dirs}{Named paths to result directories, see description.}

\item{score.column}{Which one column to read in as score.}

\item{join.column}{Which columns define an intron.}

\item{type.samples}{Forwarded to \code{IRFinderS_read()}.}

\item{min.samples}{How to filter retained introns, see description.}

\item{wl}{How to filter retained introns, see description.}
}
\value{
\code{SummarizedExperiment} object.
}
\description{
Report a "merged" SummarizedExperiment for differential analysis. Refer to
details section.
}
\details{
The routine is divided into the following steps:

First, read in retained introns from each sample by \code{IRFinderS_read}.

Introns that have warnings defined by \code{wl} is removed from each sample.
\code{wl} definition follows that of the \verb{IRFinder Diff} routine.

Unique introns are defined by columns specified in \code{join.column}.

Next, consensus introns are determined. If an intron is found in at least
\code{min.samples} of samples it is considered as a consensus intron.

Finally, intron scores from the 'full' data table are extracted,
which will be put as the "scores" assay in the output SE object.

\code{named.result.dirs} must be a named character vector whose:
\itemize{
\item names = group name
\item values = full path to the sample result directories
}
}
\examples{
#TODO
}
