% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readStarGeneCounts}
\alias{readStarGeneCounts}
\title{Read STAR Aligner Gene Counts}
\usage{
readStarGeneCounts(named_paths, clean = FALSE)
}
\arguments{
\item{named_paths}{Named list of count table paths.}

\item{clean}{Boolean, whether to remove entries starting with "N_".}
}
\value{
\code{SummarizedExperiment} of counts. Metadata \code{sample} will be names.
}
\description{
Read STAR Aligner Gene Counts
}
\examples{
# TODO
}
