% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.properties.R
\name{RIME}
\alias{RIME}
\title{Intron length ratio to mean neighboring exons}
\usage{
RIME(
  txdb = TxDb.Dmelanogaster.UCSC.dm6.ensGene::TxDb.Dmelanogaster.UCSC.dm6.ensGene,
  GRange.intron
)
}
\arguments{
\item{txdb}{Bioconductor TxDb. Must match the intron GRanges.}

\item{GRange.intron}{GenomicRanges::GRanges of the introns.}
}
\value{
Numeric vector of the ratios.
}
\description{
Intron length ratio to mean neighboring exons
}
\examples{
vignette("intron-properties")
}
