% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{rmats_filter}
\alias{rmats_filter}
\title{Assign filter labels of rMATS output}
\usage{
rmats_filter(
  df,
  readCov = 10,
  maxPSI = 0.95,
  minPSI = 0.05,
  sigFDR = 0.1,
  sigDeltaPSI = 0.05,
  res_column = "filter"
)
}
\arguments{
\item{df}{rMATS data read by \code{rmats_read()}.}

\item{readCov}{Minimum read coverage.}

\item{maxPSI}{Maximum percent spliced-in PSI.}

\item{minPSI}{Minimum PSI.}

\item{sigFDR}{FDR threshold for significance.}

\item{sigDeltaPSI}{PSI threshold for significance.}

\item{res_column}{Which column to store filter results}
}
\value{
rMATS data table with filter label column \code{filter}.
}
\description{
Assign filter labels of rMATS output
}
\details{
This function reproduces the data filter demonstrated in Xing Lab 2024
Nature Protocol tutorial. See Github repository
Xinglab/rmats-turbo-tutorial and refer to rmats_filtering.py.

AS events that does not satisfy any of the following criteria will be given
the \emph{remove} label:
\enumerate{
\item Average read counts (IJC+SJC) in both Sample 1 and Sample 2 conditions
are at least \code{readCov}.
\item Average PSI in both Sample 1 and Sample 2 conditions are between
\code{minPSI} and \code{maxPSI}.
}

If all criteria above are satisfied, events will be labeled by statistical
significance:
\enumerate{
\item \code{FDR < sigFDR} and \code{IncLevelDifference > sigDeltaPSI}: \emph{up}.
\item \code{FDR < sigFDR} and \code{IncLevelDifference < -sigDeltaPSI}: \emph{down}.
\item Otherwise: \emph{ns}.
}
}
\examples{
# TODO
}
