% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{rmats_read}
\alias{rmats_read}
\title{Read rMATS output of different AS patterns}
\usage{
rmats_read(outputs.dir, method)
}
\arguments{
\item{outputs.dir}{Output directory of rMATS.}

\item{method}{Either "JC" or "JCEC", see details.}
}
\value{
Data frame of rMATS output
}
\description{
Read rMATS output of different AS patterns
}
\details{
This is a convenience function for reading rMATS output, merging different
splicing patterns into a single data frame for easier further analysis.
Refer to rMATS Github for details on results generated by rMATS:
https://github.com/Xinglab/rmats-turbo/blob/v4.3.0/README.md#output

rMATS normalizes lengths of individual splicing variants. There are two
methods used for this normalization: JC and JCEC. Refer to the rMATS paper
for more details: https://doi.org/10.1073/pnas.1419161111

rMATS coordinated are 0-based; the exact meaning of start and end varies by
splicing pattern type:
\itemize{
\item A3SS/A5SS: start/end = start/end of the long exon (inclusion form).
\item SE: start/end = start/end of the skipped exon (inclusion form).
\item MXE: start/end = start of the first exon / end of the second exon.
\item RI: start/end = start/end of the retained intron (inclusion form).
}
}
\examples{
# TODO
}
