% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{rmats_toGRange}
\alias{rmats_toGRange}
\title{Converts rMATS data frame to GenomicRange}
\usage{
rmats_toGRange(df, ...)
}
\arguments{
\item{df}{rMATS data read by \code{rmats_read()}.}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> See details.

Three columns are used to fill in GRanges information: \code{chr}, \code{start_0base}, \code{end}.

Two columns are by default added to the metadata: \code{geneSymbol}, \code{Type}.

Extra metadata columns are specified by the tidyselect ellipsis.}
}
\value{
GRanges object
}
\description{
Convenience function for converting rMATS data to GRanges.
}
\examples{
#TODO
}
