% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcTools.R
\name{se_subset}
\alias{se_subset}
\title{Subset SummarizedExperiment}
\usage{
se_subset(se, expr_assay, expr_colData, fn_rowwise)
}
\arguments{
\item{se}{SummarizedExperiment object.}

\item{expr_assay}{Expression to evaluate on assays}

\item{expr_colData}{Expression to evaluate on colData}

\item{fn_rowwise}{Predicate to define rowwise filter}
}
\value{
Subset SummarizedExperiment object.
}
\description{
\enumerate{
\item Subset columns using \code{expr_colData}.
\item Evaluate \code{expr_assay} on \code{assays()} of the subset.
\item Apply function \code{fn_rowwise} to get subset vector.
\item Use the subset vector to subset rows of the input \code{se}.
}
}
\examples{
# TODO
}
