% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeXStringSetNamed}
\alias{writeXStringSetNamed}
\title{writeXStringSet with identifiers}
\usage{
writeXStringSetNamed(x, filepath, append = FALSE, compress = FALSE, ...)
}
\arguments{
\item{x}{Object XString to write to file}

\item{filepath}{File to write to}

\item{append}{Must be False (does not support append)}

\item{compress}{Must be False (does not support compression)}

\item{...}{Forwarded to \code{Biostrings::writeXStringSet()}}
}
\description{
Write a XStringSet to file (FASTA/FASTQ) with names as identifiers.
}
\details{
Somehow the original \code{Biostrings::writeXStringSet()} does not write
identifiers. This function uses names provided by \code{name()} and write ID.
}
\examples{
#TODO
}
