% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intron.AS.R
\name{rmats_filter}
\alias{rmats_filter}
\title{Filtering rMATS output of different AS patterns}
\usage{
rmats_filter(df, supp_reads = c(5, 2), incLvl_limits = c(0.05, 0.95))
}
\arguments{
\item{df}{Data frame of parsed rMATS output. Use \code{rmats_read()}.}

\item{supp_reads}{Supporting read filter. See details.}

\item{incLvl_limits}{Inclusion level filter. See details.}
}
\value{
Filtered rMATS output. See details.
}
\description{
Filtering rMATS output of different AS patterns
}
\details{
Typically, it is desirable to filter the rMATS output to:
\itemize{
\item Reject detected AS events that have too few supporting reads.
\item Remove AS events whose inclusion levels are extreme.
}

This function applies the following filters (default parameters assumed):
\enumerate{
\item Supporting read count (both EJC and IJC) must be >=5 in >=2 samples.
\item Inclusion level must be in the range of 0.05-0.95.
}
}
\examples{
#TODO

}
